Reconstruct force being used condition and integer state access

+++ src/legoapi/legoapi_types.h
@@ -2917,7 +2917,7 @@ struct GIZFORCE_s {
     GAMEANIMSET_s *anim_set;    // 0x28
     nuhspecial_s along_socket;  // 0x2c
     GameObject_s *using_object; // 0x38, cleared after every Force update
-    f32 field_0x3c;
+    u32 field_0x3c; // 0x3c, tested as a complete integer word by ForceBeingUsed
     GIZFORCEGROUP_s *group; // 0x40
     f32 force_strength;     // 0x44
     f32 field_0x48;
